Fix for nvidia drivers on Ubuntu 21.10 and other very modern systems
The `snapcraft-runner` script that snapcraft uses to launch apps, including sommelier-based apps, prepends `$SNAP_LIBRARY_PATH` onto `$LD_LIBRARY_PATH`. This actually needs to be addpended (on the end, not the beginning), so that the drivers are found last.

Ask @flexiondotorg for the gory details, as this is their fix discovered for the OBS-Studio snap that I've adapted ;-)